requirements.txt
# Core dependencies para Flujo_nuevo (Requerimiento → JSON)
requests>=2.28.0
numpy>=1.24.0
faiss-cpu>=1.7.0
openpyxl>=3.1.0

# PDF processing
# PyPDF2>=3.0.0
# pdfplumber>=0.11.0
# pdf2image>=1.17.0
# pytesseract>=0.3.13
# Pillow>=11.0.0
